--- src/OpenFeature/Extension/EnumExtensions.cs
@@ -1,13 +1,0 @@
-using System.ComponentModel;
-
-namespace OpenFeature.Extension;
-
-internal static class EnumExtensions
-{
-    public static string GetDescription(this Enum value)
-    {
-        var field = value.GetType().GetField(value.ToString());
-        var attribute = field?.GetCustomAttributes(typeof(DescriptionAttribute), false).FirstOrDefault() as DescriptionAttribute;
-        return attribute?.Description ?? value.ToString();
-    }
-}
