+++ src/OpenFeature/Extensions/EnumExtensions.cs
@@ -1,0 +1,13 @@
+using System.ComponentModel;
+
+namespace OpenFeature.Extensions;
+
+internal static class EnumExtensions
+{
+    public static string GetDescription(this Enum value)
+    {
+        var field = value.GetType().GetField(value.ToString());
+        var attribute = field?.GetCustomAttributes(typeof(DescriptionAttribute), false).FirstOrDefault() as DescriptionAttribute;
+        return attribute?.Description ?? value.ToString();
+    }
+}
