+++ test/OpenFeature.E2ETests/Steps/EvaluationStepDefinitions.cs
@@ -1,6 +1,6 @@
 using OpenFeature.Constant;
 using OpenFeature.E2ETests.Utils;
-using OpenFeature.Extension;
+using OpenFeature.Extensions;
 using OpenFeature.Model;
 
 namespace OpenFeature.E2ETests.Steps;
